requirements.txt
amqp==2.3.2
boto3==1.9.28
celery==4.2.1
dateparser==0.7.0
Django==2.1.2
django-allauth==0.38.0
django-compressor==2.2
django-material==1.3.0
django-model-utils==3.1.2
django-rest-auth==0.9.3
django-smtp-ssl==1.0
django-storages==1.7.1
djangorestframework==3.9.0
django-taggit==0.23.0
feedparser>5
Pillow==5.3.0
psycopg2-binary==2.7.5
readability-lxml==0.7
requests==2.20.0
uWSGI==2.0.17.1
